Moves some school idolu code around
